Move from old reporting infrastructure to metrics service
