+++ pages/downloads/routes.js
@@ -1,12 +1,6 @@
 const report = require('./reports');
-const experiments = require('./reports-experimental');
 
 module.exports = {
-  experiments: {
-    path: '/experimental/:report',
-    router: experiments,
-    breadcrumb: false
-  },
   report: {
     path: '/:report',
     router: report,
