--- pages/nts-experimental/content/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../../nts/content');
