--- pages/nts-experimental/views/index.jsx
@@ -1,3 +1,0 @@
-import Index from '../../nts/views/index';
-
-export default Index;
